server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
starlette-session==0.3.1